pyproject.toml
[project]
name = "pe-hackathon"
version = "0.1.0"
description = "URL Shortener - MLH PE Hackathon 2026"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.0",
    "peewee>=3.17.0",
    "psycopg2-binary>=2.9.9",
    "python-dotenv>=1.0.0",
    "python-json-logger>=2.0.7",
    "prometheus-client>=0.20.0",
    "psutil>=5.9.8",
    "redis>=5.0.0",
    "playhouse",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-flask>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=app --cov-report=term-missing --cov-report=xml"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
